[refactor] RacingCarMover 테스트 코드 분리
- 공통 부분 프로퍼티로 분리
- tearDownWithError로 각 테스트마다 프로퍼티 초기화
